+++ examples/examples.cmake
@@ -2,6 +2,7 @@
 if(${CMAKE_SYSTEM_NAME} STREQUAL XCORE_XS3A)
     ## Bare metal examples
     include(${CMAKE_CURRENT_LIST_DIR}/bare-metal/explorer_board/explorer_board.cmake)
+    include(${CMAKE_CURRENT_LIST_DIR}/bare-metal/visual_wake_words/visual_wake_words.cmake)
 
     ## FreeRTOS examples
     include(${CMAKE_CURRENT_LIST_DIR}/freertos/audio_mux/audio_mux.cmake)